updated ADC initialisation

+++ DSP/Mixer.cpp
@@ -3,7 +3,7 @@
 void Mixer::init(daisy::DaisySeed* seed, float* m[2], float* t1[2], float* t2[2], float* t3[2], float* t4[2], DelayLine<float, MAXDELAY>* dl[2])
 {
     bufferSize = SAMPLERATE * DURATION;
-    for(int i = 0 ; i < 2 ; i++)
+    for(uint_fast8_t i = 0 ; i < 2 ; i++)
         mix[i] = m[i];
 
     for(size_t j = 0 ; j < bufferSize ; j++)
